selection query

+++ migrations/book.sql
@@ -1,5 +1,5 @@
 create table "book"
 (
-    book_id     integer primary key,
+    book_id     bigint unsigned primary key,
     words       text    not null
 );
